Ensure local state in the current loop rather than the next one.
Previous coding would only ensure current state after a loop, meaning that
if PostgreSQL isn't running at `pg_autoctl run` start, we would wait for 5s
until we start Postgres, even after having successfully contacted the
monitor.

Update the code so that as soon as we have an answer from the Monitor with
our goal state we can ensure the current state and then move on. This allows
the Postgres service to be restarted without delay at pg_autoctl start.