Fix stale postmaster status check in pg_setup_is_ready (#2)
* Fix stale postmaster status check in pg_setup_is_ready

Previously this kept the last known state (e.g. "ready"), and if Postgres
experienced a shutdown (e.g. by another session calling "pg_ctl stop") it
would not correctly detect Postgres not running, because the state was
still seen as ready.

By invalidating our in-memory value for Postmaster status we make sure we
refresh it and see it when Postgres has been stopped.

* Ensure local state in the current loop rather than the next one.

Previous coding would only ensure current state after a loop, meaning that
if PostgreSQL isn't running at `pg_autoctl run` start, we would wait for 5s
until we start Postgres, even after having successfully contacted the
monitor.

Update the code so that as soon as we have an answer from the Monitor with
our goal state we can ensure the current state and then move on. This allows
the Postgres service to be restarted without delay at pg_autoctl start.

* Implement tests for restarting Postgres for ensuring current state.

When Postgres is stopped, the keeper process detects that fact and then
restarts Postgres when it's expected to be. This patch adds unit testing of
that facility and behavior.