Implement pg_autoctl show ipaddr|cidr.
Rather than using `hostname -fqdn` we introduce new commands `show ipaddr`
which displays the discovered LAN ip address, and `show cidr` which allows
for debugging pg_autoctl HBA edits.

Those commands would be nice additions to the PG_AUTOCTL_DEBUG mode for the
command line, but it seems that in some environments with broken reverse DNS
we need an easy way to discover the LAN IP address from the command line. A
quick search shows it's pretty hard to do that with classic OS tooling in a
portable way, and we already have had to write code for that feature in
pg_auto_failover anyway. This patch exposes the feature to the end users.

See #6.