Use the default --nodename discovery in the test suite.
We only do that for the Postgres node, not for the monitor. The monitor
discovery of the --nodename involves a connection attempt to external
services (Google public DNS service at 8.8.8.8) and we don't want the test
suite to fail when that's not reachable from the local network setup.

The Postgres node connection attempt is made targetting the monitor address,
which certainly should be reachable, and done with the proper outbound local
IP address.