Implement checking and upgrading of the pgautofailover extension.
The pg_autoctl tool embeds the version of the extension that it is
compatible with on the monitor. The idea is that those versions are kept in
sync at the source code and packaging level, and a given pg_autoctl command
line is compatible with a single version of the monitor.

At startup on the monitor the command `pg_autoctl run` checks the monitor's
version and if necessary runs a command `ALTER EXTENSION ... UPDATE TO ...`
the expected version. It is expected that the extension packaging provides
the necessary upgrade paths.

On the keeper, we check that the version number of the extension on the
monitor is compatible with our expectations, and exit before entering the
main loop when that's not the case. We refrain from doing "upgrade at a
distance" and tell the admin (who's reading the logs, maybe?) to please take
care of the situation on the monitor node.