Use --nodename when adding local network CIDR range to HBA.
We used to try to discover the local IP address to use, but now that we have
a default --nodename, we know we have a config.nodename at this step of the
setup. So use the provided or discovered nodename and open the CIDR of this
item in the monitor HBA rules by default.

In passing, the given or discovered nodename might be the IP address of a
loopback interface, which we must not reject then. The user might know what
they are doing, and that includes our unit test environment.