Implement extension version testing in the binary.
The SQL schema for the extension includes catalog metadata of the C coded
functions we provide. Those definitions must be in sync with the actual code
being loaded. The code embeds its own version information and can query
Postgres catalogs to get the version numbers for the extension.

In this patch we add a test at each exposed SQL function that the current
version of the extension loaded as shared library has the same version
number as the currently installed extension (otherwise restarting Postgres
is needed), and the same version as the currently available extension
version (otherwise ALTER EXTENSION UPDATE is needed).