Manually set LANG=C when running pg_controldata. (#27)
We parse the output of pg_controldata and expect English output. Force the
environment to be in C locale when calling pg_controldata so that we do not
depend on the user locale settings here.

The other output parsing we do is on pg_ctl --version, where we only parse
the version string. As that's only digits and dots ([[:digit:].]+), we don't
need to set the LANG=C in that case. I think. At least this patch does
nothing for this case.

Fix #26.