Implement pg_autoctl enable|disable maintenance CLI.
The feature was already there and the CLI hidden in our DEBUG tooling,
because we were not sure at the time of the initial implementation how to
expose the feature to our users.

Now that we have the enable|disable subcommands already, it looks like it is
a good fit for the maintenance feature.

Fixes #33.