Fix dockerfile build-dependencies.
Now that we're using $(shell $(PG_CONFIG) --libs) to have the list of
librairies we depend on, we must align the build dependencies of
pg_auto_failover with the Postgres build dependencies on the platform where
we build.

In debian, that means having build dependencies towards readline
support (that's libedit), pam, zlib, xml, xstl, and selinux.

See #34.