Fix memory context when performing version check
String buffer returned by TextDatumGetCString was in
SPI memory context, and it is invalidated upon SPI_finish()

We set active memory context here so that data will continue
to be available after SPI_finish(). It would be reclaimed
by next memory context reset to occur after completion of the call.