Release ipdb and iproute to avoid resource exhaustion in tests (#226)
DESCRIPTION: fix an issue in the test suite that was limiting the amount of virtual nodes that could be created.

The test suite had a flaw where only 55 virtual nodes could be created. This caused added tests to fail 'randomly' while succeeding in isolation. By correctly closing both the uses of IPDB and IPRoute we can now create upto 168 virtual nodes before running into the same flaw again.

It is unclear what the current limitation is, but this gives us enough headroom for the foreseeable future to keep adding tests.

Special thanks to Hadi for finding the missing release calls.