Implement support for PostgreSQL 12 "standby.signal".
In Postgres 12 the setup to start as a standby has changed. The standby
configuration parameters go in the main postgresql.conf file, and the
decision to start in standby mode depends on the existence of a new file,
named "standby.signal".

Adapting the code to this new system requires some refactoring.