WIP: command listener process.
We review our process hierarchy so that pg_autoctl may be the entry point
for managing the whole Postgres processes, and offer some HTTPd service in
parallel to that.

If the HTTPd service runs the pg_autoctl and postgres service itself, then
we have the HTTP port open in all the process hierarchy, which is better
avoided. Instead, implement a listener process that can take commands over a
pipe channel.