Add a `do fsm check` command and run it once in a while.
The new command ensures that the current state is actually what our FSM
expects, making sure that Postgres is running when it is supposed to run,
etc.

We call into that command every 5s from the http process so that there's no
risk of concurrency issues (race conditions) with the other http users. We
only implement one HTTP action at a time, it's all sync, on purpose.