Fix creating a secondary server in an already existing directory.
This could have been a one-liner, but raises the question of the hard-coded
choice for the backup directory. The major reason of the target PGDATA
already existing is that it's targeting a mount-point. Then our default
choice of ${PGDATA}/../backup for the pg_basebackup target is not a good
one, because the renaming of the directory will cross file system boundaries
and not be atomic anymore.

This patch then allows to configure the backup directory. In passing, the
default value now includes the nodename, so that it's easier to run tests
environments with multiple instances on the same machine.