Fix execv() failure handling and argv0 usage.
In some cases:

  - we would use argv[0] as the binary to invoke again in sub-programs
  - execv() would fail with "No such file or directory"

Then our handling of execv() failure was leaving a process and a pipe
behind, leading to a strange process tree and a bad situation.

This patch fixes the code so that we get the realpath for the currently
executed program (using proper API in macOS but a heuristic for other OSes),
and fixes the error handling at the execv() call site too.