Implement JSON format for HTTP route /api/1.0/state.
For that we implement a new --json option to the command "show state", and
use it from the HTTP route. Building the JSON output is left to Postgres
thanks to jsonb_pretty() and other json building functions.

In passing, make sure to run the HTTP *and* the keeper listener service when
running with a monitor. In that case, as we are also running the node active
protocol process, we don't need the HTTP service to run checks for us, so we
disable them.