Implement pg_autoctl restart safe cache for config and hba files.
In case pg_autoctl (or even the whole node) is restarted in the middle of a
transition, we might find ourselves unable to set the file paths cache from
the live Postgres server. That happens when we got restarted at a point
where Postgres should NOT be running, so we keep it that way.

In that case, use the values we have in our pg_autoctl configuration file,
that's our cache for those settings. In a real life installation changing
the location of the config_file and the hba_file “never happens”… unless
maybe when using debian packaging tool pg_createcluster on a pre-existing
Postgres cluster.