Tolerate missing optional INI options in missing sections. (#109)
Without this patch when reading the configuration file and it's missing a
section entirely, it would be an error even for optional parameters. This
situation mostly happens when upgrading pg_autoctl and the new version has
implemented a new configuration section, which didn't exist previously, and
has default parameters that you can run with.