WIP: fix pg_setup_standby_mode.
When introducting support for Postgres 12 we also introduced a bug when
installing the standby.signal file expects a running Postgres to get the
file path for the HBA setup... at a time where it's required that Postgres
is not running.

The fix consists of caching the file path locations of postgresql.conf and
pg_hba.conf. Those are not expected to ever change, and we still invalidate
the cache at each keeper loop, to be sure (see pg_createcluster in debian
packaging tools: it could still happen).