Simplify the Postgres metadata fetching logic.
We went up from 2 to 4 queries in this PR, where it could have been done in
a single query. And since we're fetching the current LSN, we can use a smart
query that works both for a primary and a standby, so that we can simplify
the code even more.