More cleanup and bugfixes, and ensure postgres is running at times.
It used to be that we would ensure the current state before making the
transition. In the general case that's the wrong thing to do. If the new
state is "demoted" you don't want to ensure the current state "primary".

Now that we only ensure the current state when there's no transition to be
implemented, we then have cases where we need to ensure bits of the expected
state in the transition code.