Implement a principled way to check fi we can run without a monitor.
Every command needs to now figure out for itself whether it's compatible
with having the monitor disabled. Rather than patch every command in an
extensive way, which is prone to error, add a monitorDisabledIsOk to the
keeper_config_read_file() calls.

In passing, rename the missingPgdataIsOk and pgIsNotRunningIsOk to their
camelCase version, which is the coding standard in Citus. We don't always
use that standard, because the function that calls pg_ctl start is better
off with being named pg_ctl_start than pgCtlStart, I believe. Still, for
those variable names, it's always been a stretch.