Feature/disable monitor (#113)
* Implement starting the service with monitor disabled.

Not having a monitor has some impact in the internals of the code, which
assumes we need to have the monitor URI and be able to contact it. This
patch has some refactoring to stop assuming we should have a monitor to talk
to, and makes it so that `pg_autoctl run` only starts the HTTPd service.

We're missing proper signal handling in the HTTPd service, so you need to
either kill -9 or make it crash for it to stop, won't answer to C-c,
somehow. That's for the next patch in the series though.

Every command needs to now figure out for itself whether it's compatible
with having the monitor disabled. Rather than patch every command in an
extensive way, which is prone to error, add a monitorDisabledIsOk to the
keeper_config_read_file() calls.

In passing, rename the missingPgdataIsOk and pgIsNotRunningIsOk to their
camelCase version, which is the coding standard in Citus. We don't always
use that standard, because the function that calls pg_ctl start is better
off with being named pg_ctl_start than pgCtlStart, I believe. Still, for
those variable names, it's always been a stretch.