Late night attempt at fixes.
Turns out we didn't handle the case when currentLSN and sync_state are NULL
in the output for the query we run to get fresh Postgres metadata, and the
way the query is built it may happen.

Also, we used to care that in SECONDARY_STATE or CATCHINGUP_STATE we would
have a current LSN on the replica, and this got lost in the refactoring.
Reinstall that. If that works, we should also re-install the error message
when we've lost the primary.