When going from primary to single, drop all replication slots.
It used to be that we could have only one replication slot with an known
name, because the name was hardcoded. Now that we can have multiple standby
servers, we don't know the name of the replication slots to drop, nor how
many there are. But we know we want to drop all of them, so just ask
Postgres to do that for us.