Create and Advance replication slots on all the standby servers.
Our standby might need to help each other out in case of primary outage, for
that to happen we maintain replication slots for every standby node on every
other standby node.

We implement the create-or-advance logic in SQL in this patch, because that
makes the code simpler. We might want to revisit that later, though... let's
say I like it that way.