Fix initialiazing an already existing Postgres cluster. (#115)
Fix initialiazing an already existing Postgres cluster.

We refuse to own an already running Postgres cluster. Then it's fair for us
to also accept that the user might just stop the Postgres instance and give
it to us again. This patch makes us recognize the situation and proceed with
our initialisation steps in that case.