Add default log level to Postgres setup.
For the monitor, because of the way it's typically setup, we need to
implement our prefered logging setup. For the keeper Postgres nodes, it is
still possible to override our choices the usual ways, so it looks okay to
provide with a default that allows better debug facilities for us.