Implement pg_autoctl show nodes.
This is like `pg_autoctl do monitor get others` and shares the same output
code and foramts (tabular and JSON), but shows every node in a given
formation, or given formation and group.

This looks like a useful command for the users to debug their setup when
using multiple standbys and so it should probably not be hidden in the debug
command line with `pg_autoctl do monitor get ...`.