Prepare for dynamic synchronous_standby_names setting.
The synchronous_standby_names setting is how we implement our policy of
number_sync_standbys, candidate priorities, and replication quorum. That's
where all of those notions are implemented in Postgres configuration.

In this patch we introduce a function that computes the setting value on the
monitor depending on the current values of our parameters. This function is
only called in the new subcommand pg_autoctl show synchronous_standby_names
at the moment though.