Fix synchronous_standby_names recipe.
We want all the nodes with replicationQuorum set to true to participate in
the setting of synchronous_standby_names, even when they have a candidate
Priority set to zero.

The priority is used for the failover and means that nodes with a priority
of zero will never be the target of the failover. They might still
participate in the replication quorum.

This is useful for an architecture with a node that can be trusted with the
data but not with the main production service (shared standby node, though
it could also be a pg_walreceiver connection actually).