Fix/error handling improvements (#133)
* After a WARNing when things get back to okay, inform the user.

In this case it's when we've lost the connection to a Postgres server and we
keep trying to reconnect, it's better to tell the user when we've been able
to connect again.

* Improve main loop logs when something goes wrong and is okay again.

We issue a WARN message when something failed in our tests, wait for 5s, and
try again. Now if we succeed we should mention it, otherwise it's impossible
to reconstruct the past states of the nodes just by reading through the logs.

* Improve SQL queries error messages.

Split the Postgres message on a line-by-line basis for better integration in
pg_autoctl logs / output, also add query parameters in case of query failure.

* Desultory code cleanup.

* Reduce some chatter (INFO message to DEBUG).

* Attempt at fixing a SEGFAULT.

* Cleanup.

* Improve multi-line error output.

We can use the new code for PQerrorMessage in more places, that makes for a
way cleaner output from pg_autoctl. More places to cover later.

* Add default log level to Postgres setup.

For the monitor, because of the way it's typically setup, we need to
implement our prefered logging setup. For the keeper Postgres nodes, it is
still possible to override our choices the usual ways, so it looks okay to
provide with a default that allows better debug facilities for us.

* Typo fix.