Include more changes from the main PR.
The cut/split was not done properly, for the tests to pass we need a little
bit more infrastructure in this patch already. It has to do with how to find
the primary node in a group when there's an unknown number of servers, where
previously all we needed to think about was "the primary is the other node".