Fix/test suite (#139)
* Tests clean-up and simplification.

* Stop testing as soon as we hit a failure.

Our testing suite is written in a way that each test depend on the previous
one being successful. That's fine. Just tell it to the test driver.

* Improve the Makefile targets for tests.

Allow to have a TEST variable to run limited tests when necessary.

* Improve tests output.

In case of failure to reach the assigned state, we want to know more. Add
the pg_autoctl process logs and the monitor events.

* Ensure we read from the pg_autoctl output pipes.

Otherwise we stall the process and it stops working. And then our tests fail
dependning on how much output has been sent before, our process stopping
when the pipe is full (typically 4kb if I remember correctly).

* Small fix: only stop existing processes.

* Tests review / indenting / verbosity.

* Refactor testing code for better code re-use.

Now that we have the PGAutoCtl class, add an "execute" method to it,
allowing more use cases and a nice refactoring of the test cases.

* Refactor pg_autoctl some more.

Make it a single place where we lookup for the pg_autoctl executable in the
PATH in the source code.