Send a SIGQUIT to pg_autoctl in the background to stop it.
We used to send a SIGTERM and this doesn't guarantee prompt stop of the
pg_autoctl process. In cases, we even hit the 60s timeout for commands,
which smokes of a loop in pg_autoctl that doesn't check if SIGTERM was
received.

Well, we actually want to stop pg_autoctl now, so send SIGQUIT.