Fix/cli perform destroy (#141)
* Add support for perform_failover() at the CLI level.

In passing, review the pg_autoctl destroy command and make it an option to
the pg_autoctl drop node command.

* Make perform_failover a SECURITY DEFINER function.

This makes it possible for autoctl_node to call the function and see the
expected result: a failover is happening.

Fixes #94.

* Send a SIGQUIT to pg_autoctl in the background to stop it.

We used to send a SIGTERM and this doesn't guarantee prompt stop of the
pg_autoctl process. In cases, we even hit the 60s timeout for commands,
which smokes of a loop in pg_autoctl that doesn't check if SIGTERM was
received.

Well, we actually want to stop pg_autoctl now, so send SIGQUIT.