Feature/multiple other node (#134)
* Allow a third Postgres server to join a group.

This makes it an extra secondary server, and for that we transition from
PRIMARY to the new state JOIN_PRIMARY. This means we still have a secondary
to failover to should the need appear, and we are concurrently adding a
second standby server to our group.

This is all still very much Work In Progress, as I don't expect anything
other than registration at the right timing to kind of work.

We still have to fix our usage of the replication slot name from the
primary, where it doesn't make sense anymore. Also we have to create
replication slots on every other secondary now. And so much more.

* Restrict standby registration to one-at-a-time.

Given our current FSM implementation and protocol it's impossible to know
which standby is free to reach CATCHING UP when the primary has prepared the
replication settings (HBA and replication slot). So rather than figuring out
a way to extend the protocol to accept registering several standby at a
time, we implement a retryable error on the monitor.

And we also implement a retry-loop on the client side. Which is implemented
as a tail-call recursive call at the moment, because that's easy.

* Return the LSN and isPrimary? in pgautofailover.get_other_nodes.

We will need this information to be able to create and manager replication
slots on each standby. When using multiple standbys we want to make sure we
keep the WAL we got for any other standby, in case we lose the primary. To
prepare for being able to do that, return the information we need from the
call to get_other_nodes.

In passing, add a JSON output to the `pg_autoctl do show others` command.
That makes it easy to programmatically parse the output, if needed. And then
we can display a human oriented table output for the command when --json is
not used.

* Fix how backupDirectory is computed, use nodeId.

This allows to have many nodes with the same nodename (typically "localhost"
when testing) and let them have each their own backup directory by default.

* Implement pg_autoctl show nodes.

This is like `pg_autoctl do monitor get others` and shares the same output
code and formats (tabular and JSON), but shows every node in a given
formation, or given formation and group.

This looks like a useful command for the users to debug their setup when
using multiple standbys and so it should probably not be hidden in the debug
command line with `pg_autoctl do monitor get ...`.