Refactoring and fixes.
Review more of the code and fix some missing parts and oddities. In
particular the following 3 API entry points didn't get the memo that we
don't have an "otherNode" that makes sense anymore:

  - start_maintenance
  - stop_maintenance
  - remove_node

In passing review and adjust some of the decision making around counting
candidate nodes and selecting a candidate: we don't want to failover to a
node that is currently in maintenance, for instance.