Per review, implement a timeout to the notification waiting loops.
It might be that we will miss the notifications we are interested into.
Usually, it takes up to 5s for the pg_autoctl background process to contact
the monitor and realize there's a state change to apply_settings, and then
up to a couple of seconds to do the whole 4 messages dance.

We implement a generous 30s timeout so that in case we miss the
notifications from the monitor, we user is not stuck indefinitely. After all
the user might be a non-interactive script.