Review get/set commands and implement set node nodename (#146)
* Clean-up the get/set commands CLI.

Make it so that each of the set properties supported in the get/set commands
are a sub-command, which provides a better user-experience and integration.

* Introduce new state APPLY_SETTINGS.

Restrict changing the replication settings of an existing formation to a
stable situation when we have a node in the "primary" state. In that case
only we can go forward with the replication changes, through the new state
"apply_settings".

* Improve pg_autoctl enable|disable maintenance.

Rather than signaling the pg_autoctl process in the hope that it reaches the
next state faster, implement a notification handler than watches until we
know the target state has been reported.

* Implement pg_autoctl set node nodename.

That updates the nodename on the monitor and then in the local configuration
file too.

* Per review, implement a timeout to the notification waiting loops.

It might be that we will miss the notifications we are interested into.
Usually, it takes up to 5s for the pg_autoctl background process to contact
the monitor and realize there's a state change to apply_settings, and then
up to a couple of seconds to do the whole 4 messages dance.

We implement a generous 30s timeout so that in case we miss the
notifications from the monitor, we user is not stuck indefinitely. After all
the user might be a non-interactive script.