Implement pg_autoctl show uri --json.
And make it so that by default we output a table of the URIs for both the
monitor and every formation that exists on the monitor. That's more useful
to the users, some of them being confused as to how to benefit from their
nice Postgres setup once the pg_autoctl create commands have done their
magic.