Fix/json (#149)
* Directly prints the JSON string.

When passing a static buffer we might run short of bytes to host the whole
JSON string: it turns out that the JSON format for the monitor state for
only 4 nodes is already well more than 1024 bytes.

* Implement pg_autoctl show events --json.

* Implement JSON format support for pg_autoctl show uri.

When using the JSON format, we output both the monitor and the formation
connection strings.