Remove pgsql_get_config_file_path.
Instead, use the following piece of code where we consider that the config
file has to be found in $PGDATA. This is a requirement of pg_auto_failover
anyway because otherwise pg_basebackup fails to create a standby node.

+	/* configFilePath = $PGDATA/postgresql.conf */
+	join_path_components(configFilePath, pgSetup->pgdata, "postgresql.conf");

Another PR is being cooked that forces the situation where the config file,
hba file and ident file are found (or moved) in $PGDATA when giving a
Postgres cluster to pg_auto_failover.