Introduce new state JOIN_SECONDARY.
When going from REPORT_LSN to SECONDARY, we need to wait until the failover
has happened on the candidate. Also, in the case where the primary would be
on the losing side of a network split, we want to make sure every replica
has stopped its replication stream with the primary. We do that by stopping
Postgres when in JOIN_SECONDARY state.

Then as soon as the candidate has reached a primary state (WAIT_PRIMARY or
PRIMARY) then we can edit our recovery settings (primary_conninfo) and start
Postgres again.