When promoting a new primary, prepare replication (HBA) for all nodes.
Before this patch we would only target the nodes that are currently in the
DEMOTED TIMEOUT state. Handling multiple standby that may each go at their
own rythm through more steps in the FSM makes it hard to prepare the HBA in
this “Just in Time” fashion. Instead, prepare replication for every node
that is currently registered in the system.

In passing, add a test that perform a failover with a standby being put in
MAINTENANCE at the same time.