Maintain replication slots for the primary too.
Each standby node maintains a replication slot for every other node in the
system, taking their current LSN position from the monitor. The lag is not
crucial here, as the goal is to ensure that we have enough of the WAL bytes
still available for everyone after a failover.

We now maintain such a replication slot for the primary node, because in
case of failover, it's going to be another secondary and have to re-attach
to the new primary. It's best if we can ensure that the new primary still
has enough WAL to make it possible to rewind or just even continue streaming
without interruption.