Refrain from dropping replication slots during failover.
We used to maintain replication slots on only standby servers, so at
promotion time after a rewind we would drop the replication slot, and when
disabling synchronous replication too.

Removing replication slots prevents from some an outage where the WAL disk
gets full while the other node fails to re-join the group.

That said, by keeping the replication slot around at failover times, we can
guarantee that the node will be able to continue replicating when it's back
without having to pg_basebackup again from scratch.

It is still possible to remove the replication slot and force a full copy of
the data to take place by dropping the missing node on the monitor.