Refrain from maintaining replication slots on Postgres 10.
The function pg_replication_slot_advance() appeared in Postgres 11. This
means that our approach where we maintain replication slots for every other
node is not possible to implement with Postgres 10.

It might be that we can ship without the strong guarantee that WALs are kept
around in every standby. We might have to implement the pg_rewind and
pg_basebackup in more places for Postgres 10, but let's first try to see if
the tests are happy that way.