Improve our recovery config clean-up.
In Postgres 12 pg_basebackup achieve setting recovery configuration by
editing the postgresql.auto.conf file. We need to edit those settings away
when we want to change the recovery setup. We can use ALTER SYSTEM RESET
when Postgres is running. It's not always the case though, so when Postgres
is not running we rewrite the postgresql.auto.conf file and skip through
the lines with the recovery settings.