Review the fast-forward recovery wait loop.
We don't want to wait until end-of-recovery, because we now pause at
recovery target. Instead I wanted to rely on pg_is_wal_replay_paused() but
found that it would always return false in our case. So we're now parsing
the pg_last_wal_replay_lsn() until it is the same as our target LSN, and
then we CHECKPOINT and finish the transition to FAST_FORWARD.