Use application_name in primary_conninfo and fix related bugs (#154)
* Use application_name in our primary_conninfo setting.

When we set synchronous_standby_names to a subset of the existing standby
nodes, we need to list they by name. The names we use are going to be
checked against the application_name in the replication connection string.

* Introduce pg_autoctl show synchronous_standby_names.

* Remove pgsql_get_config_file_path.

Instead, use the following piece of code where we consider that the config
file has to be found in $PGDATA. This is a requirement of pg_auto_failover
anyway because otherwise pg_basebackup fails to create a standby node.

+	/* configFilePath = $PGDATA/postgresql.conf */
+	join_path_components(configFilePath, pgSetup->pgdata, "postgresql.conf");

* When using Postgres 12, pg_basebackup uses postgresql.auto.conf.

This means that our include file postgresql-auto-failover-standby.conf can
contains any value we want, they won't be used if postgresql.auto.conf has a
setting in place. That's particulary inconvenient for primary_conninfo.

So we now do ALTER SYSTEM RESET primary_conninfo; when going from REPORT_LSN
to JOIN_SECONDARY, so that then when we go to SECONDARY we can follow the
new primary and have Postgres use our new setup.