Review the wal cascade test case and the group state machine.
Rather than using the perform failover API, inject a real node failure in
the test case. That way, we don't have to have the nodes in a stable state
to begin with.

Then, fix the Group FSM to handle some of the missing transitions when nodes
appear in the middle or after the failover is done. In particular, that
means reviewing how we find which node is currently the primary in the
group. When a DEMOTED node joins back in the middle of a failover, it might
not be as obvious as one might like.