Review debian test cases.
Make it so that we use --listen in the debian environement, so that we use a
pghost connection that reaches our HBA edits. The debian HBA default entries
would forbid our attempt to connect to localhost.

In passing, refrain from accepting some errors with enabling the local LAN
CIDR in the test environment. We certainly want to see those.