Quick & dirty fix for a bootstrap issue with replication password.
We can't set the pg_autoctl configuration file entry for the replication
password until the configuration file has been created, which happens in the
pg_autoctl create command. That said, pg_basebackup fails in the create
command because the replication password is not set yet.

The dirty fix is to do it twice and set the replication password in between
the two calls.

A better fix would be to introduce a way to break the circular dependency:

  - provide a way to set the replication password at create time,
  - or separate the registration step from the pg_basebackup step.

The second approach is more appealing because it avoids yet another password
handling mechanism to implement, or passing it in the clear.