Implement an option to skip editing of HBA rules by pg_autoctl (#169)
* Add forgotten --auth method to keeper's config.

The setting should be found in the configuration file and re-installed in
memory at pg_autoctl start-up, like it is the case for the monitor. This
looks like it was forgotten and bugs never actually showed-up.

* Fix a segmentation fault hazard.

When an option is not set (replication.password in this testing session),
don't try to dereference its strValue pointer: it's NULL.

* Improve "silent" error logging.

In cases we hide the logs to the users, and skip some ERROR messages. It
happens that sometimes we want to see the exact same error messages at the
DEBUG level in those cases.

* Implement --skip-pg-hba.

This is implemented using a new authentication method: "skip". This allows
minimal changes to the internals of the code, and to issue a nice WARNing
message to our users so that they can easily figure out what changes they
need to make to their HBA files.

* Improve pg_autoctl config check.

Include testing of connection strings, allowing for better debugging of HBA
rules. That's going to be useful for --skip-pg-hba users.

* Require --auth trust rather than making it the silent default.

* Improve log output of subcommands.

* Add a test case for --skip-pg-hba.

This is work in progress and allows to test placing one's own SSL
certificates and setup and HBA rules when bypassing the default authMethod
provided by pg_autoctl usually.

* Now use sslmode=verify-ca in the test for --skip-pg-hba.

* Implement certificate based authentication in --skip-pg-hba test.

This time we use the full SSL settings in a way that we have authentication
without passwords.

* Update Python version on Travis.

* Update documentation with --auth and --skip-pg-hba options.

Also update some more that we missed in previous rounds.

* Quick & dirty fix for a bootstrap issue with replication password.

We can't set the pg_autoctl configuration file entry for the replication
password until the configuration file has been created, which happens in the
pg_autoctl create command. That said, pg_basebackup fails in the create
command because the replication password is not set yet.

The dirty fix is to do it twice and set the replication password in between
the two calls.

A better fix would be to introduce a way to break the circular dependency:

  - provide a way to set the replication password at create time,
  - or separate the registration step from the pg_basebackup step.

The second approach is more appealing because it avoids yet another password
handling mechanism to implement, or passing it in the clear.

* Apply review comments.

* Test that the HBA files are not editing when using --skip-pg-hba.

Weel actually... our test environment *requires* some changes to the HBA, so
even when using --skip-pg-hba the HBA file is edited to add two rules. Take
that into account in the test suite thanks to diff --ignore-matching-lines.