Implement sslMode for replication connection strings.
We need to support sslmode=... in both the pg_basebackup command line when
creating a secondary, and when installing our connection string in the
primary_conninfo setting in either recovery.conf or our own configuration
file.

The monitor connection string is a copy/paste by the user, and so all we can
do is make sure we output the proper connection string with an sslmode
parameter in the monitor's output.