Review the SSL options to force users to be explicit.
We now have --no-ssl and --ssl-self-signed and the per-file options when
certificates are user-provided. There's no default value, much like --auth,
so the users must choose the option they prefer and be concious about it.

The sslmode is now preserved in the configuration and our connection strings
are using the sslmode as setup. The default value for sslmode depends on the
SSL settings. When using verify-ca or verify-full, we use --ssl-ca-file as
the sslrrotcert and --ssl-crl-file as the sslcrl connection parameters.