Implement support for SSL connections (#171)
In this PR we implement SSL support for all pg_auto_failover connections.
There are multiple ways to do it, from self-signed certificates to full
blown CA signed certificates with authentication.

We introduce new options to handle the 3 server-side SSL settings that we
support, plus the client-side SSL settings for building our connection
strings.

We need to support sslmode=... in both the pg_basebackup command line when
creating a secondary, and when installing our connection string in the
primary_conninfo setting in either recovery.conf or our own configuration
file.

The monitor connection string is a copy/paste by the user, and so all we can
do is make sure we output the proper connection string with an sslmode
parameter in the monitor's output.