Use our own secure version of fprintf.
There are several security issues with fprintf that we want to avoid:
  - the FILE * stream might be NULL,
  - the format string might be NULL,
  - any %s argument might be NULL.

What happens in those cases is undefined, so we'd rather make sure those
cases are not happening. This PR introduces pgaf_fprintf() where we add the
first two checks, and then import pg_fprintf() from Postgres port directory
to fix the third issue.