Implement sformat, a secured version of snprintf. (#186)
* Implement sformat, a secured version of snprintf.

Replace all our calls to snprintf with sformat, which is based on Postgres
pg_snprintf code and adds a couple of checks against NULL strings.

* Log about BUGgy situations in fformat and sformat.

It's ok for us to expect that in most call sites of sformat the expanded
format string is going to fit in the given buffer (of BUFSIZE = 1024 bytes),
though when it does not we should at least log_error about the situation.