Fix replication slot handling.
Replication slots now have a name that includes their consumer node id (the
standby, typically). This means that we need to know the name of id of the
other node when dropping its replication slot, or when going single, that we
need to drop all the existing replication slots, rather than the one with a
static name.