Use a postgres sub-process in pg_autoctl create monitor.
This means that when not using the --run option, when the pg_autoctl command
is finished, Postgres is not running. It's an important change from before,
and is expected.

That way there will be no confusion about what is the monitor process,
either pg_autoctl or postgres. Now, the monitor's Postgres instance is
running within the pg_autoctl run command.