Fix shutdown sequence.
Use killpg(2) instead of kill(2) in pg_autoctl stop, so that each one of the
pg_autoctl services (sub-processes) that have been started receive the
signal, rather than only the process leader itself.

In passing, in case people would use kill(1) or pkill(1), detect that the
sub-processes are still running after a while (~5s in this commit) and send
the SIGTERM signal to the process group manually.