Per review, refrain from using pg_basebackup --write-recovery-conf.
This option adds the replication settings to postgresql.auto.conf which is
trouble some for us. Instead of coping with that fact by editing the auto
configuration file even when Postgres is not running, we just bypass the
feature entirely from pg_basebackup and use our own routines to create the
setup.

We need our own routines anyway in other places.