Feature/replication slots (#222)
* When going from primary to single, drop all replication slots.

It used to be that we could have only one replication slot with an known
name, because the name was hardcoded. Now that we can have multiple standby
servers, we don't know the name of the replication slots to drop, nor how
many there are. But we know we want to drop all of them, so just ask
Postgres to do that for us.

* Create and Advance replication slots on all the standby servers.

Our standby might need to help each other out in case of primary outage, for
that to happen we maintain replication slots for every standby node on every
other standby node.

We implement the create-or-advance logic in SQL in this patch, because that
makes the code simpler.

* Implement slot maintenance on the single/primary node.

* Review replication slots handling in several cases.

In particular review the way creating a replication slot can be made
idempotent. It used to be that we would drop the slot and create it again.

In the case of multiple standby we might have failed to create one of
several slots, so we want to create-if-not-exists instead. That's what we do
now.

In passing, secure the dropping of "all" slots to be limited to physical
slots with a name that follows pg_auto_failover naming pattern. After all
other replication slots might be there that we don't own.

* Maintain replication slots for the primary too.

Each standby node maintains a replication slot for every other node in the
system, taking their current LSN position from the monitor. The lag is not
crucial here, as the goal is to ensure that we have enough of the WAL bytes
still available for everyone after a failover.

We now maintain such a replication slot for the primary node, because in
case of failover, it's going to be another secondary and have to re-attach
to the new primary. It's best if we can ensure that the new primary still
has enough WAL to make it possible to rewind or just even continue streaming
without interruption.

* Refrain from dropping replication slots during failover.

We used to maintain replication slots on only standby servers, so at
promotion time after a rewind we would drop the replication slot, and when
disabling synchronous replication too.

Removing replication slots prevents from an outage where the WAL disk
gets full while the other node fails to re-join the group.

That said, by keeping the replication slot around at failover times, we can
guarantee that the node will be able to continue replicating when it's back
without having to pg_basebackup again from scratch.

It is still possible to remove the replication slot and force a full copy of
the data to take place by dropping the missing node on the monitor.

* Refrain from maintaining replication slots too early.

Now that every node maintains replication slots for every other node in the
group, we want to wait until the monitor assigns SECONDARY state before
creating the slots: it means we have reached an LSN position where it makes
it possible.

* Review tests and ensure both pg_autoctl processes are making progress.

We need to make sure that all the pg_autoctl running processes are not
prevented from making progress: someone needs to read from those PIPEs.

* Improve our recovery config clean-up.

In Postgres 12 pg_basebackup achieve setting recovery configuration by
editing the postgresql.auto.conf file. We need to edit those settings away
when we want to change the recovery setup. We can use ALTER SYSTEM RESET
when Postgres is running. It's not always the case though, so when Postgres
is not running we rewrite the postgresql.auto.conf file and skip through
the lines with the recovery settings.

* Per review, refrain from using pg_basebackup --write-recovery-conf.

This option adds the replication settings to postgresql.auto.conf which is
trouble some for us. Instead of coping with that fact by editing the auto
configuration file even when Postgres is not running, we just bypass the
feature entirely from pg_basebackup and use our own routines to create the
setup.

We need our own routines anyway in other places.

Co-authored-by: Jelte Fennema <github-tech@jeltef.nl>